move secondsToHumanReadable to utils
